refactor: deprecate ModelFactory
