+++ src/Bundle/Maker/Factory/FactoryClassMap.php
@@ -12,7 +12,7 @@
 namespace Zenstruck\Foundry\Bundle\Maker\Factory;
 
 use Zenstruck\Foundry\Bundle\Maker\Factory\Exception\FactoryClassAlreadyExistException;
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 
 /**
  * @internal
@@ -24,14 +24,14 @@ final class FactoryClassMap
      */
     private array $classesWithFactories;
 
-    /** @param \Traversable<ModelFactory> $factories */
+    /** @param \Traversable<PersistentProxyObjectFactory> $factories */
     public function __construct(\Traversable $factories)
     {
         /** @phpstan-ignore-next-line */
         $this->classesWithFactories = \array_unique(
             \array_reduce(
                 \iterator_to_array($factories, preserve_keys: true),
-                static function(array $carry, ModelFactory $factory): array {
+                static function(array $carry, PersistentProxyObjectFactory $factory): array {
                     $carry[$factory::class] = $factory::getEntityClass();
 
                     return $carry;
