+++ tests/Fixtures/Factories/AddressFactory.php
@@ -11,13 +11,13 @@
 
 namespace Zenstruck\Foundry\Tests\Fixtures\Factories;
 
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Address;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-final class AddressFactory extends ModelFactory
+final class AddressFactory extends PersistentProxyObjectFactory
 {
     protected static function getClass(): string
     {
