+++ tests/Fixtures/Factories/ContactFactory.php
@@ -11,13 +11,13 @@
 
 namespace Zenstruck\Foundry\Tests\Fixtures\Factories;
 
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Contact;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-final class ContactFactory extends ModelFactory
+final class ContactFactory extends PersistentProxyObjectFactory
 {
     protected static function getClass(): string
     {
