+++ tests/Fixtures/Factories/ODM/CommentFactory.php
@@ -11,11 +11,11 @@
 
 namespace Zenstruck\Foundry\Tests\Fixtures\Factories\ODM;
 
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Tests\Fixtures\Document\ODMComment;
 use Zenstruck\Foundry\Tests\Fixtures\Document\ODMUser;
 
-class CommentFactory extends ModelFactory
+class CommentFactory extends PersistentProxyObjectFactory
 {
     protected static function getClass(): string
     {
