+++ tests/Fixtures/Factories/TagFactory.php
@@ -11,13 +11,13 @@
 
 namespace Zenstruck\Foundry\Tests\Fixtures\Factories;
 
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Tag;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-final class TagFactory extends ModelFactory
+final class TagFactory extends PersistentProxyObjectFactory
 {
     protected static function getClass(): string
     {
