+++ tests/Fixtures/Maker/expected/can_create_factory_for_entity_with_repository_with_data_set_psalm.php
@@ -11,14 +11,14 @@
 
 namespace App\Factory;
 
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Proxy;
 use Zenstruck\Foundry\RepositoryProxy;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Post;
 use Zenstruck\Foundry\Tests\Fixtures\Repository\PostRepository;
 
 /**
- * @extends ModelFactory<Post>
+ * @extends PersistentProxyObjectFactory<Post>
  *
  * @method        Post|Proxy                     create(array|callable $attributes = [])
  * @method static Post|Proxy                     createOne(array $attributes = [])
@@ -52,7 +52,7 @@ use Zenstruck\Foundry\Tests\Fixtures\Repository\PostRepository;
  * @psalm-method static list<Proxy<Post>> randomRange(int $min, int $max, array $attributes = [])
  * @psalm-method static list<Proxy<Post>> randomSet(int $number, array $attributes = [])
  */
-final class PostFactory extends ModelFactory
+final class PostFactory extends PersistentProxyObjectFactory
 {
     /**
      * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#factories-as-services
