+++ tests/Fixtures/Maker/expected/can_create_factory_for_not_persisted_class_interactively.php
@@ -11,19 +11,19 @@
 
 namespace App\Factory;
 
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Proxy;
 use Zenstruck\Foundry\Tests\Fixtures\Object\SomeObject;
 
 /**
- * @extends ModelFactory<SomeObject>
+ * @extends PersistentProxyObjectFactory<SomeObject>
  *
  * @method        SomeObject|Proxy     create(array|callable $attributes = [])
  * @method static SomeObject|Proxy     createOne(array $attributes = [])
  * @method static SomeObject[]|Proxy[] createMany(int $number, array|callable $attributes = [])
  * @method static SomeObject[]|Proxy[] createSequence(iterable|callable $sequence)
  */
-final class SomeObjectFactory extends ModelFactory
+final class SomeObjectFactory extends PersistentProxyObjectFactory
 {
     /**
      * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#factories-as-services
