+++ tests/Fixtures/Maker/expected/can_create_factory_for_odm_with_data_set_embedded_document.php
@@ -11,20 +11,20 @@
 
 namespace App\Factory;
 
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Proxy;
 use Zenstruck\Foundry\Tests\Fixtures\Document\ODMComment;
 use Zenstruck\Foundry\Tests\Fixtures\Factories\ODM\UserFactory;
 
 /**
- * @extends ModelFactory<ODMComment>
+ * @extends PersistentProxyObjectFactory<ODMComment>
  *
  * @method        ODMComment|Proxy     create(array|callable $attributes = [])
  * @method static ODMComment|Proxy     createOne(array $attributes = [])
  * @method static ODMComment[]|Proxy[] createMany(int $number, array|callable $attributes = [])
  * @method static ODMComment[]|Proxy[] createSequence(iterable|callable $sequence)
  */
-final class ODMCommentFactory extends ModelFactory
+final class ODMCommentFactory extends PersistentProxyObjectFactory
 {
     /**
      * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#factories-as-services
