+++ tests/Fixtures/Maker/expected/can_create_factory_in_test_dir.php
@@ -12,13 +12,13 @@
 namespace App\Tests\Factory;
 
 use Doctrine\ORM\EntityRepository;
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Proxy;
 use Zenstruck\Foundry\RepositoryProxy;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Category;
 
 /**
- * @extends ModelFactory<Category>
+ * @extends PersistentProxyObjectFactory<Category>
  *
  * @method        Category|Proxy                   create(array|callable $attributes = [])
  * @method static Category|Proxy                   createOne(array $attributes = [])
@@ -36,7 +36,7 @@ use Zenstruck\Foundry\Tests\Fixtures\Entity\Category;
  * @method static Category[]|Proxy[]               randomRange(int $min, int $max, array $attributes = [])
  * @method static Category[]|Proxy[]               randomSet(int $number, array $attributes = [])
  */
-final class CategoryFactory extends ModelFactory
+final class CategoryFactory extends PersistentProxyObjectFactory
 {
     /**
      * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#factories-as-services
