+++ tests/Fixtures/Maker/expected/can_create_factory_in_test_dir_interactively.php
@@ -12,13 +12,13 @@
 namespace App\Tests\Factory;
 
 use Doctrine\ORM\EntityRepository;
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Proxy;
 use Zenstruck\Foundry\RepositoryProxy;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Tag;
 
 /**
- * @extends ModelFactory<Tag>
+ * @extends PersistentProxyObjectFactory<Tag>
  *
  * @method        Tag|Proxy                        create(array|callable $attributes = [])
  * @method static Tag|Proxy                        createOne(array $attributes = [])
@@ -36,7 +36,7 @@ use Zenstruck\Foundry\Tests\Fixtures\Entity\Tag;
  * @method static Tag[]|Proxy[]                    randomRange(int $min, int $max, array $attributes = [])
  * @method static Tag[]|Proxy[]                    randomSet(int $number, array $attributes = [])
  */
-final class TagFactory extends ModelFactory
+final class TagFactory extends PersistentProxyObjectFactory
 {
     /**
      * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#factories-as-services
