+++ tests/Fixtures/Maker/expected/can_create_factory_with_default_enum_with_data_set_odm.php
@@ -12,14 +12,14 @@
 namespace App\Factory;
 
 use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Proxy;
 use Zenstruck\Foundry\RepositoryProxy;
 use Zenstruck\Foundry\Tests\Fixtures\PHP81\DocumentWithEnum;
 use Zenstruck\Foundry\Tests\Fixtures\PHP81\SomeEnum;
 
 /**
- * @extends ModelFactory<DocumentWithEnum>
+ * @extends PersistentProxyObjectFactory<DocumentWithEnum>
  *
  * @method        DocumentWithEnum|Proxy             create(array|callable $attributes = [])
  * @method static DocumentWithEnum|Proxy             createOne(array $attributes = [])
@@ -37,7 +37,7 @@ use Zenstruck\Foundry\Tests\Fixtures\PHP81\SomeEnum;
  * @method static DocumentWithEnum[]|Proxy[]         randomRange(int $min, int $max, array $attributes = [])
  * @method static DocumentWithEnum[]|Proxy[]         randomSet(int $number, array $attributes = [])
  */
-final class DocumentWithEnumFactory extends ModelFactory
+final class DocumentWithEnumFactory extends PersistentProxyObjectFactory
 {
     /**
      * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#factories-as-services
