+++ tests/Fixtures/Maker/expected/can_create_factory_with_embeddable_with_data_set_orm.php
@@ -12,13 +12,13 @@
 namespace App\Factory;
 
 use Doctrine\ORM\EntityRepository;
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Proxy;
 use Zenstruck\Foundry\RepositoryProxy;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Contact;
 
 /**
- * @extends ModelFactory<Contact>
+ * @extends PersistentProxyObjectFactory<Contact>
  *
  * @method        Contact|Proxy                    create(array|callable $attributes = [])
  * @method static Contact|Proxy                    createOne(array $attributes = [])
@@ -36,7 +36,7 @@ use Zenstruck\Foundry\Tests\Fixtures\Entity\Contact;
  * @method static Contact[]|Proxy[]                randomRange(int $min, int $max, array $attributes = [])
  * @method static Contact[]|Proxy[]                randomSet(int $number, array $attributes = [])
  */
-final class ContactFactory extends ModelFactory
+final class ContactFactory extends PersistentProxyObjectFactory
 {
     /**
      * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#factories-as-services
