+++ tests/Fixtures/Object/SomeOtherObjectFactory.php
@@ -11,18 +11,18 @@
 
 namespace Zenstruck\Foundry\Tests\Fixtures\Object;
 
-use Zenstruck\Foundry\ModelFactory;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Proxy;
 
 /**
- * @extends ModelFactory<SomeOtherObject>
+ * @extends PersistentProxyObjectFactory<SomeOtherObject>
  *
  * @method        SomeOtherObject|Proxy     create(array|callable $attributes = [])
  * @method static SomeOtherObject|Proxy     createOne(array $attributes = [])
  * @method static SomeOtherObject[]|Proxy[] createMany(int $number, array|callable $attributes = [])
  * @method static SomeOtherObject[]|Proxy[] createSequence(iterable|callable $sequence)
  */
-final class SomeOtherObjectFactory extends ModelFactory
+final class SomeOtherObjectFactory extends PersistentProxyObjectFactory
 {
     protected function getDefaults(): array
     {
